catch blur events and attempt to use first suggestion if present
This is a sort of poor-man's autocomplete allowing for fuzzy matching at
both the beginning and end of the line while still allowing the user to
continue typing after a match (ergo not having to use a key delay).

The basic idea is from the [SO answer from 4nti][source], but the
implementation is a little different. As the popup still exists when the
blur event gets called, we don't have to save the value on every key
event. On the other hand the popup doesn't actually have the value we
need to set the widget to, it has the label value. Parsing this back to
the value involves doing a lookup in the store and some other stuff I
couldn't figure out, so I cheated. Emitting a click on the popup item
seems to take care of all that at the expense of needing two tab events
to actually get out of the field.

This isn't perferct, but it's an improvement.

  [source]: http://stackoverflow.com/a/30715425/313192